requirements.txt
customtkinter
nltk
pillow
pyinstaller
pygame
google-cloud-texttospeech
elevenlabs
pydub
requests
packaging